+++ utils/brookesia_lib_utils/src/time_profiler.cpp
@@ -24,7 +24,7 @@ void TimeProfiler::set_format_options(const FormatOptions &options)
     BROOKESIA_LOG_TRACE_GUARD_WITH_THIS();
 
     BROOKESIA_LOGI(
-        "Set format options: %1%", BROOKESIA_DESCRIBE_TO_STR_WITH_FMT(options, detail::DESCRIBE_FORMAT_VERBOSE)
+        "Set format options: %1%", BROOKESIA_DESCRIBE_TO_STR_WITH_FMT(options, DESCRIBE_FORMAT_VERBOSE)
     );
 
     std::lock_guard<std::mutex> lock(mutex_);