+++ utils/brookesia_lib_utils/CHANGELOG.md
@@ -1,6 +1,19 @@
 # ChangeLog
 
-### Initial Release:
+### v0.7.1 - 2025-12-05
+
+#### Enhancements:
+
+* feat(describe_helpers): Add describe helpers for converting between any type and string (JSON format), JSON and any type
+* feat(describe_helpers): Add support for 'std::variant', 'std::function'
+* feat(state_machine): Add 'wait_all_transitions()', 'force_transition_to()' methods
+* feat(task_scheduler): Add 'dispatch()' method
+
+#### Bug Fixes:
+
+* fix(log): Fix garbled display issue for 'int8_t'
+
+### v0.7.0 - 2025-11-28
 
 - feat(repo): Add task scheduler (based on Boost.Asio) supporting multi-threaded task management, including immediate, delayed, and periodic tasks
 - feat(repo): Support thread configuration (name, priority, stack size, stack location, CPU binding configuration)